--- backend/was/src/common/decorators/swagger/tarot.decorator.ts
@@ -1,18 +1,0 @@
-import { createFindByDecorator } from './factory/decorator.factory';
-import { Param } from './interface';
-
-export const FindTarotCardDecorator = (
-  target: string,
-  param: Param,
-  returnType: any,
-) => {
-  return createFindByDecorator(target, returnType, param);
-};
-
-export const FindTarotResultDecorator = (
-  target: string,
-  param: Param,
-  returnType: any,
-) => {
-  return createFindByDecorator(target, returnType, param);
-};
